Update useNotifications with new useAnalytics
